+++ src/ics/CalendarPlugin.php
@@ -42,6 +42,12 @@ class CalendarPlugin extends AbstractPlugin
                 }
                 return $this->routeHandler;
             },
+            'freebusy' => function($authService) {
+                if ($this->routeHandler === null) {
+                    $this->routeHandler = new CalendarRouteHandler($authService);
+                }
+                return $this->routeHandler;
+            },
             'caldav' => function($authService) {
                 if ($this->caldavRouteHandler === null) {
                     $this->caldavRouteHandler = new CalDAVRouteHandler($authService);